版主转让以及小bug的修复

+++ src/main/java/com/techwave/entity/dto/TransferSectionDTO.java
@@ -13,8 +13,8 @@ import lombok.NoArgsConstructor;
 
 public class TransferSectionDTO {
 
-    private Integer targetId;
+    private Long targetId;
 
-    private Integer sectionId;
+    private Long sectionId;
 
 }
